% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCounts.R
\name{cleanCounts}
\alias{cleanCounts}
\title{Remove effects from covariates of interest}
\usage{
cleanCounts(object, effectNames, byNames = NULL)
}
\arguments{
\item{object}{an epigraHMMDataSet}

\item{effectNames}{a character vector with the names of assays for which the effect will be removed from the experimental counts.
Names in `effectNames` must be assays stored in the epigraHMMDataSet `object`.}

\item{byNames}{a character vector with the name of an assay containing stratification variables which will be used to define stratum-specific effects.
Examples of byNames assays include the 'peaks' assay from `initializer()`. In this case, models will be fit separately for peaks and non-peaks regions.
This can be usefor for effects such as GC content, which are known to have a differential effect between peaks and non-peak regions.
Default is NULL, i.e., effects will be removed without stratification.}
}
\value{
An epigraHMMDataSet with an 'offset' assay filled in.
}
\description{
This function removes the effect from covariates of interest (such as GC content) from experimental counts
}
\details{
To be added
}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
