% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBam.R
\name{getBam}
\alias{getBam}
\title{Download and rename bam files from chipseqDBData}
\usage{
getBam(dir = tempdir(), download = FALSE)
}
\arguments{
\item{dir}{a (temporary) path where bam files will be stored}

\item{download}{a logical indicating whether or not files should be downloaded}
}
\value{
A GRanges object with differential peak calls in BED 6+3 format
}
\description{
This function downloads and renames bam files from chipseqDBData to be used 
in the vignette of epigraHMM
}
\examples{

getBam(download = FALSE)

}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
