% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalize counts}
\usage{
normalizeCounts(object, type = "loess", ...)
}
\arguments{
\item{object}{an epigraHMMDataSet}

\item{type}{either 'loess', the default, or 'gam'}

\item{...}{arguments to be passed to \code{\link[mgcv]{gam}} or \code{\link[limma]{loessFit}} for loess calculation}
}
\value{
An epigraHMMDataSet with an 'offsets' assay filled in.
}
\description{
This function performs a non-linear normalization of counts with respect to a reference sample (geometric mean)
}
\details{
If `type = 'gam'`, `normalizeCounts` will use `mgcv::bam` with option `discrete = TRUE`.
If `type = 'loess'`, it will use `limma::loessFit`, which simply a wrapper for the `stats::lowess` smoother.
}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
