% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatterns.R
\name{plotPatterns}
\alias{plotPatterns}
\title{Create a plot of differerential patterns posterior probabilities from epigraHMM}
\usage{
plotPatterns(object, ranges, hdf5, colors = NULL)
}
\arguments{
\item{object}{an epigraHMMDataSet}

\item{ranges}{a GRanges object or a pair of integers with the genomic corrdinates/windows to be plotted}

\item{hdf5}{a character string with the hdf5 file path from `epigraHMM`}

\item{colors}{an optional argument that specifies the colors for each differential combinatorial pattern}
}
\value{
A pheatmat
}
\description{
`plotPatterns()` plots the posterior probabilities associated with differential patterns from a differential analysis of `epigraHMM()`
}
\examples{
# Creating dummy object
countData <- cbind(rbind(matrix(rnbinom(1e2,mu = 2,size = 10),ncol = 1),
                         matrix(rnbinom(1e2,mu = 7.5,size = 5),ncol = 1),
                         matrix(rnbinom(1e2,mu = 7.5,size = 5),ncol = 1),
                         matrix(rnbinom(1e2,mu = 7.5,size = 5),ncol = 1),
                         matrix(rnbinom(1e2,mu = 2,size = 10),ncol = 1)),
                   rbind(matrix(rnbinom(1e2,mu = 2,size = 10),ncol = 1),
                         matrix(rnbinom(1e2,mu = 7.5,size = 5),ncol = 1),
                         matrix(rnbinom(1e2,mu = 7.5,size = 5),ncol = 1),
                         matrix(rnbinom(1e2,mu = 2,size = 10),ncol = 1),
                         matrix(rnbinom(1e2,mu = 2,size = 10),ncol = 1)))

colData <- data.frame(condition = c('A','B'), replicate = c(1,1))
rowRanges <- GRanges('chrA',
                     IRanges(start = seq(1,by = 500,length.out = nrow(countData)),
                             width = 500))

object <- epigraHMMDataSetFromMatrix(countData,colData,rowRanges = rowRanges)

# Initializing
object <- initializer(object,controlEM())

# Running epigraHMM
object <- epigraHMM(object,controlEM(),type = 'differential',dist = 'nb')

# Calling peaks
peaks <- callPeaks(object = object,
                  hdf5 = S4Vectors::metadata(object)$output,
                  method = 'viterbi')

# Plotting patterns
plotPatterns(object,peaks,S4Vectors::metadata(object)$output)

}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
