% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateMarkovChain}
\alias{simulateMarkovChain}
\title{Simulates a Markov Chain of length 'n' given a matrix of transition probabilities P}
\usage{
simulateMarkovChain(P, n)
}
\arguments{
\item{P}{a matrix of transition probabilities (row sums should be 1)}

\item{n}{an integer specifying thhe length of the simulated sequence}
}
\description{
Simulates a Markov Chain of length 'n' given a matrix of transition probabilities P
}
\examples{
#Example
simulateMarkovChain(matrix(c(0.99,0.01,0.01,0.99),2,2),100)
}
