% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOffsets.R
\name{addOffsets}
\alias{addOffsets}
\title{Add offsets to epigraHMMDataSet}
\usage{
addOffsets(object, offsets)
}
\arguments{
\item{object}{an epigraHMMDataSet}

\item{offsets}{a matrix with model offsets}
}
\value{
An epigraHMMDataSet with an 'offsets' assay filled in.
}
\description{
This function adds model offsets to epigraHMMDataSet
}
\details{
To be added
}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
}
