% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraHMM.R
\name{epigraHMM}
\alias{epigraHMM}
\title{Perform peak calling of epigenomic data sets}
\usage{
epigraHMM(object, control, type, dist, random = FALSE)
}
\arguments{
\item{object}{an epigraHMMDataSet}

\item{control}{list of control arguments from \code{\link{controlEM}}}

\item{type}{character, either \code{"consensus"} or \code{"differential"}}

\item{dist}{character, either \code{"zinb"} or \code{"nb"}}
}
\value{
To be added
}
\description{
This function runs either consensus (one condition, multiple samples) or
differential (multiple conditions and samples) peak callers for epigenomic
data.
}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
