% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraHMMDataSetFromBam.R
\name{epigraHMMDataSetFromBam}
\alias{epigraHMMDataSetFromBam}
\title{Create a epigraHMMDataSet from a set of BAM files}
\usage{
epigraHMMDataSetFromBam(
  bamFiles,
  colData,
  genome,
  windowSize,
  gapTrack = TRUE,
  blackList = TRUE,
  param = readParam()
)
}
\arguments{
\item{bamFiles}{a string vector (or a list of string vectors) with the path for BAM files. If bamFiles is a list of string vectors,
vectors must be named, have the same dimension, and, at least, a vector with name 'counts' must exist (see details).}

\item{colData}{a \code{data.frame} with the experimental data. It must contain the columns \code{condition} and \code{replicate}.
\code{condition} refers to the experimental condition identifier (e.g. cell line name). \code{replicate} refers to the replicate identification number (unique for each condition).}

\item{genome}{either a single string with the name of the reference genome (e.g. 'hg19') or a GRanges object with ranges to be tilled into a set of non-overlapping windows.}

\item{windowSize}{an integer specifying the size of genomic windows where read counts will be computed.}

\item{gapTrack}{either a logical (\code{TRUE}, the default, or \code{FALSE}) or a GRanges object with gap regions of the genome to be excluded. If \code{TRUE}, the function will discard genomic coordinates overlapping regions present in the UCSC gap table of the respective reference genome (if available). See Details section below.}

\item{blackList}{either a logical (\code{TRUE}, the default, or \code{FALSE}) or a GRanges object with blacklisted regions of the genome to be excluded. If \code{TRUE}, the function will discard ENCODE blacklisted regions from selected reference genomes (if available). See Details section below.}

\item{param}{an object from the \code{\link[csaw]{readParam}} class with specifications for the read counting step.}
}
\value{
An epigraHMMDataSet object with sorted colData regarding conditions and replicates.
Experimental counts will be stored in the 'counts' assay in the resulting epigraHMMDataSet object.
Additional experimental data will be stored with their respective names from the list bamFiles.
}
\description{
This function creates a \code{\link[SummarizedExperiment]{RangedSummarizedExperiment}} object from of a set of BAM files.
It is used to store the input data, the model offsets, and the results from the peak calling algorithms.
}
\details{
The index ".bai" files must be stored in the same directory of their respective BAM files.
The index files must be named after their respective BAM files with the additional ".bai" suffix.

`epigraHMMDataSetFromBam` will store experimental data (e.g. ChIP-seq counts) from bamFiles (or bamFiles[['counts']], if a list is provided).
Additional data (e.g. input control counts) will be stored similarly with their respective list names.

By default, the function computes read counts using csaw's estimated fragment length via cross correlation analysis.
For experimental counts (e.g. ChIP-seq), sequencing reads are extended (if single-end reads) and scaled to the geometric mean of the average lengths of the sequenced fragments in each library.
For additional counts (e.g. input control), sequencing reads are extended to the geometric mean of the of the average lengths of the experimental libraries.

Additional columns included in the colData input will be passed to the resulting epigraHMMDataSet assay and can be acessed via \code{colData()} function.

The \code{genome} argument will call Seqinfo::Seqinfo() to fetch the chromosome lengths of the specified genome.
See ?Seqinfo::Seqinfo for the list of UCSC genomes that are currently supported.

If \code{gapTrack = TRUE} and the name of a reference genome is passed as input through \code{genome} (e.g. 'hg19'),
the function will discard any genomic coordinate overlapping regions specified by the respective UCSC gap table.
If \code{gapTrack} is a GRanges object, the function will discard any genomic coordinate overlaping regions from \code{gapTrack}.

If \code{blackList = TRUE} and the name of a reference genome is passed as input through \code{genome} (e.g. 'hg19'),
The function will fetch the manually curated blacklist tracks (Version 2) from \url{https://github.com/Boyle-Lab/Blacklist/tree/master/lists}.
Current available genomes are ce10, dm3, hg19, hg38, and mm10.
If \code{blackList} is a GRanges object, the function will discard any genomic coordinate overlaping regions from \code{blackList}.
}
\examples{
bamFiles <- system.file("extdata",
                        "wgEncodeBroadHistoneH1hescCtcfStdAlnRep1.chr21.bam",
                        package="genomationData")

colData <- data.frame(condition = 'CTCF', replicate = 1)

object <- epigraHMMDataSetFromBam(bamFiles = bamFiles,
                                  colData = colData,
                                  genome = 'hg19',
                                  windowSize = 25000,
                                  gapTrack = TRUE,
                                  blackList = TRUE)

}
\references{
\url{https://github.com/plbaldoni/epigraHMM}
DOI: 10.1093/nar/gkv1191
DOI: 10.1038/s41598-019-45839-z
DOI: 10.1038/nature11247
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
